+++ packages/atlas/src/.env
@@ -4,6 +4,7 @@ VITE_ENV=development
 VITE_GEOLOCATION_SERVICE_URL=https://geolocation.joystream.org
 VITE_PRICE_SERVICE_URL=
 VITE_ASSETS_SERVICE_URL=https://atlas-services.joystream.org/avatars
+VITE_HCAPTCHA_SITE_KEY=69bf583a-6e4b-4272-9d2e-8c4337e25187
 
 # default target env is development
 VITE_DEVELOPMENT_ORION_URL=https://atlas-dev.joystream.org/orion/graphql
