+++ packages/atlas/src/api/client/index.ts
@@ -3,7 +3,7 @@ import { BatchHttpLink } from '@apollo/client/link/batch-http'
 import { WebSocketLink } from '@apollo/client/link/ws'
 import { getMainDefinition } from '@apollo/client/utilities'
 
-import { ORION_GRAPHQL_URL, QUERY_NODE_GRAPHQL_SUBSCRIPTION_URL } from '@/config/urls'
+import { ORION_GRAPHQL_URL, QUERY_NODE_GRAPHQL_SUBSCRIPTION_URL } from '@/config/env'
 
 import cache from './cache'
