+++ packages/atlas/src/components/ViewErrorFallback/ViewErrorFallback.tsx
@@ -6,8 +6,8 @@ import { useNavigate } from 'react-router-dom'
 import { AnimatedError } from '@/components/AnimatedError'
 import { Text } from '@/components/Text'
 import { Button } from '@/components/_buttons/Button'
+import { JOYSTREAM_DISCORD_URL } from '@/config/env'
 import { absoluteRoutes } from '@/config/routes'
-import { JOYSTREAM_DISCORD_URL } from '@/config/urls'
 import { media, sizes } from '@/styles'
 import { SentryLogger } from '@/utils/logs'
