+++ packages/atlas/src/components/_auth/SignInModal/SignInModal.tsx
@@ -5,8 +5,8 @@ import shallow from 'zustand/shallow'
 
 import { Button } from '@/components/_buttons/Button'
 import { DialogButtonProps } from '@/components/_overlays/Dialog'
+import { FAUCET_URL } from '@/config/env'
 import { JOY_CURRENCY_TICKER } from '@/config/joystream'
-import { FAUCET_URL } from '@/config/urls'
 import { MemberId } from '@/joystream-lib'
 import { hapiBnToTokenNumber } from '@/joystream-lib/utils'
 import { useJoystream } from '@/providers/joystream'
