+++ packages/atlas/src/components/_overlays/AdminModal/AdminModal.tsx
@@ -12,9 +12,9 @@ import { Select } from '@/components/_inputs/Select'
 import { Switch } from '@/components/_inputs/Switch'
 import { DialogModal } from '@/components/_overlays/DialogModal'
 import { availableNodes } from '@/config/availableNodes'
+import { NODE_URL } from '@/config/env'
 import { BUILD_ENV, availableEnvs } from '@/config/envs'
 import { absoluteRoutes } from '@/config/routes'
-import { NODE_URL } from '@/config/urls'
 import { useConfirmationModal } from '@/providers/confirmationModal'
 import { useEnvironmentStore } from '@/providers/environment'
 import { useSnackbar } from '@/providers/snackbars'
