+++ packages/atlas/src/components/_overlays/TransactionModal/TransactionModal.tsx
@@ -4,7 +4,7 @@ import { CSSTransition } from 'react-transition-group'
 import { LottiePlayer } from '@/components/LottiePlayer'
 import { Text } from '@/components/Text'
 import { SvgActionCheck, SvgLogoPolkadot } from '@/components/_icons'
-import { JOYSTREAM_STORAGE_DISCORD_URL } from '@/config/urls'
+import { JOYSTREAM_STORAGE_DISCORD_URL } from '@/config/env'
 import { ErrorCode, ExtrinsicStatus } from '@/joystream-lib'
 import { useOverlayManager } from '@/providers/overlayManager'
 import { useUser, useUserStore } from '@/providers/user'
