+++ packages/atlas/src/config/env.ts
@@ -22,3 +22,4 @@ export const ATLAS_GITHUB_URL = 'https://github.com/Joystream/atlas'
 
 export const JOY_PRICE_SERVICE_URL = readEnv('PRICE_SERVICE_URL', false, true)
 export const USER_LOCATION_SERVICE_URL = readEnv('GEOLOCATION_SERVICE_URL', true, true)
+export const HCAPTCHA_SITE_KEY = readEnv('HCAPTCHA_SITE_KEY', false, true)
