+++ packages/atlas/src/main.tsx
@@ -3,8 +3,8 @@ import { createRoot } from 'react-dom/client'
 import { AssetLogger, SentryLogger } from '@/utils/logs'
 
 import { App } from './App'
+import { ASSET_LOGS_URL, SENTRY_DSN } from './config/env'
 import { BUILD_ENV } from './config/envs'
-import { ASSET_LOGS_URL, SENTRY_DSN } from './config/urls'
 
 const initApp = async () => {
   if (BUILD_ENV === 'production') {
