+++ packages/atlas/src/providers/assets/assets.manager.tsx
@@ -3,8 +3,8 @@ import { FC, useEffect } from 'react'
 
 import { StorageDataObjectFieldsFragment } from '@/api/queries'
 import { ASSET_RESPONSE_TIMEOUT } from '@/config/assets'
+import { DISTRIBUTOR_ASSET_PATH } from '@/config/env'
 import { BUILD_ENV } from '@/config/envs'
-import { DISTRIBUTOR_ASSET_PATH } from '@/config/urls'
 import { useUserLocationStore } from '@/providers/userLocation'
 import { joinUrlFragments } from '@/utils/asset'
 import { AssetLogger, ConsoleLogger, DataObjectResponseMetric, DistributorEventEntry, SentryLogger } from '@/utils/logs'
