+++ packages/atlas/src/providers/assets/assets.provider.tsx
@@ -27,8 +27,8 @@ import {
 } from '@/api/queries'
 import { ViewErrorFallback } from '@/components/ViewErrorFallback'
 import { ASSET_MIN_DISTRIBUTOR_REFETCH_TIME } from '@/config/assets'
+import { USER_LOCATION_SERVICE_URL } from '@/config/env'
 import { absoluteRoutes } from '@/config/routes'
-import { USER_LOCATION_SERVICE_URL } from '@/config/urls'
 import { useMountEffect } from '@/hooks/useMountEffect'
 import { UserCoordinates, useUserLocationStore } from '@/providers/userLocation'
 import { ConsoleLogger, SentryLogger } from '@/utils/logs'
