+++ packages/atlas/src/providers/joystream/joystream.provider.tsx
@@ -2,8 +2,8 @@ import BN from 'bn.js'
 import { ProxyMarked, Remote, proxy, wrap } from 'comlink'
 import { FC, PropsWithChildren, createContext, useCallback, useEffect, useRef, useState } from 'react'
 
+import { JOY_PRICE_SERVICE_URL, NODE_URL } from '@/config/env'
 import { JOY_CURRENCY_TICKER } from '@/config/joystream'
-import { JOY_PRICE_SERVICE_URL, NODE_URL } from '@/config/urls'
 import { JoystreamLib } from '@/joystream-lib'
 import { HAPI_TO_JOY_RATE } from '@/joystream-lib/config'
 import { useEnvironmentStore } from '@/providers/environment/store'
