+++ packages/atlas/src/views/playground/Playgrounds/PlaygroundCaptcha.tsx
@@ -2,27 +2,20 @@ import HCaptcha from '@hcaptcha/react-hcaptcha'
 import { useState } from 'react'
 
 import { Text } from '@/components/Text'
-import { BUILD_ENV } from '@/config/envs'
-
-// TODO: this must be kept in some safe place
-const DEVELOPMENT_SITE_KEY = '10000000-ffff-ffff-ffff-000000000001'
-const PRODUCTION_SITE_KEY = '150fc9ad-4a29-405e-9711-6571f91a0aeb'
+import { HCAPTCHA_SITE_KEY } from '@/config/env'
+import { ConsoleLogger } from '@/utils/logs'
 
 export const PlaygroundCaptcha = () => {
   const [response, setResponse] = useState<string>()
   const handleVerify = async (response: string) => {
+    ConsoleLogger.log(response)
     setResponse(response)
   }
 
   return (
     <div>
       {!response ? (
-        <HCaptcha
-          sitekey={BUILD_ENV === 'production' ? PRODUCTION_SITE_KEY : DEVELOPMENT_SITE_KEY}
-          theme="dark"
-          languageOverride="en"
-          onVerify={handleVerify}
-        />
+        <HCaptcha sitekey={HCAPTCHA_SITE_KEY} theme="dark" languageOverride="en" onVerify={handleVerify} />
       ) : (
         <Text variant="h500" as="p" margin={{ bottom: 10 }}>
           Captcha solved!
